requirements.txt
RPi.GPIO
spidev
mysql-connector-python